feat(rust): use crate `prelude::*` for single source of truth for imports, remove wildcard exports (#9704)
* chore(rust): Refactor error variant and match arm generation logic

Deduplicates API error variants by name and groups errors by status code for improved match arm generation. Standard error variants are now built via a helper, and semantic fields are mapped directly by status code. Multi-error match arms use a discriminator for error type, and field extraction logic is streamlined for both string and u64 types. This refactor improves maintainability and correctness of error handling code generation.

* chore(rust): parameter methods to accept Into<Option<T>>

Updated int, float, and bool parameter methods in QueryBuilder to accept impl Into<Option<T>> for more flexible argument types. Also made minor formatting improvements and removed unnecessary trait bounds.

* chore(rust): Improve Rust type generation and trait derivation logic

Refactored Rust model generators to more accurately derive PartialEq, Hash, and Eq traits based on field types. Updated primitive and container type handling for trait support checks, improved import logic for collections and BigInt, and fixed Rust type mapping for integer and float primitives. This enhances correctness and flexibility in generated Rust code for complex models.

* chore(rust): Rust SDK client generation for subpackages

Improves client file generation logic for subpackages by generating unified mod.rs files for nested subclients and skipping separate client files in those cases. Refines re-export logic, enhances import analysis for required types, and updates query parameter serialization to handle Option wrapping. This change ensures correct module structure and import usage for complex package hierarchies.

* chore(rust): Rust codegen for named type exports

Improves module and type export logic in Rust generators by ensuring unique module/type exports, applying consistent snake_case formatting, and switching to named exports instead of wildcard exports. Refactors import analysis in SubClientGenerator to only include direct parameter types, reducing unnecessary imports for serialized types. Also adds logic to export only top-level subpackage clients in SDK generator.

* feat(rust): Add prelude module and use it for imports in Rust SDK

Introduces a new `prelude.rs` module that re-exports commonly used types and traits for convenient imports. Updates code generation and documentation snippets to use the prelude for all crate imports, simplifying import statements and improving ergonomics. Also refactors type utility functions for stricter TypeScript typing and improves error handling logic.

* chore(rust): client struct and constructor to use AST

Replaces manual string-based construction of the client struct and its constructor with Rust AST classes for fields, struct, and impl block. This improves maintainability and leverages the codegen library for generating Rust code.

* chore(rust): update snapshots

Updated struct and enum fields in Rust model and SDK test snapshots to use i64 instead of i32 for integer types. Also added PartialEq, Eq, and Hash derives to several types for improved comparison and hashing. Fixed Option<String> mapping in SDK error parsing logic to avoid unnecessary allocation.

* Automated update of seed files

* Automated update of seed files

* chore(rust): bump version 0.5.2 with import fixes

Introduces version 0.5.2 to the Rust SDK. This update uses crate prelude::* for a single source of truth for imports and removes wildcard exports, improving code clarity and maintainability.

---------

+++ seed/rust-model/client-side-params/src/search_resources_request.rs
@@ -1,12 +1,12 @@
 use std::collections::HashMap;
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 pub struct SearchResourcesRequest {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub query: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub filters: Option<HashMap<String, serde_json::Value>>,
-    pub limit: i32,
-    pub offset: i32,
+    pub limit: i64,
+    pub offset: i64,
 }
