+++ seed/rust-sdk/pagination/src/api/types/inline_users_inline_users_list_users_pagination_response.rs
@@ -10,6 +10,6 @@ pub struct ListUsersPaginationResponse {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub page: Option<Page>,
     /// The totall number of /users
-    pub total_count: i32,
+    pub total_count: i64,
     pub data: Users,
 }
