+++ seed/rust-sdk/server-sent-event-examples/src/api/types/completions_streamed_completion.rs
@@ -4,5 +4,5 @@ use serde::{Deserialize, Serialize};
 pub struct StreamedCompletion {
     pub delta: String,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub tokens: Option<i32>,
+    pub tokens: Option<i64>,
 }
