+++ seed/rust-sdk/streaming/src/api/types/generate_stream_request.rs
@@ -3,5 +3,5 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Hash)]
 pub struct GenerateStreamRequest {
     pub stream: bool,
-    pub num_events: i32,
+    pub num_events: i64,
 }
